Add --export-unsafe option for tool authors
The ability to ask for unsafe declarations to be exported is useful for
creating test data or counter-examples. It's much easier and less
error-prone to write these in Lean as unsafe declarations and export
them as such than it is to try and write them from primitives in an
external tool.