chore: Use ExceptT for error handling
As a triage to #8, adds some non-panic handling for the case where
a constant to be dumped by name is not found in the environment. This
should probably be considered an unrecoverable error in almost all
scenarios, so it simply throws and terminates with an error message
stating what declaration wasn't found, then terminates with a non-zero
error code (1).

For the other disallowed export items (mvar, fvar, mdata for now) which
are currently marked `unreachable`, they might as well be handled with
a little bit more information and similarly fail immediately.